spb_hackaton/backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
aiosqlite
python-dotenv
python-multipart
